[JD-13]Refactor: 게시물, 댓글 타입 재사용 가능하게 따로 작성
[JD-13]Refactor: 게시물, 댓글 타입 재사용 가능하게 따로 작성
